test: added and updated tests to reflect new worlds to world flow
